CI/Cache: [dune cache] is job specific and pull-push
We enforce that dune cache will always have the job name in its key and will pull the cache at start and pushes the cache at the end of the job.

This simplifies the code and reflects the current usage of dune caches. These changes may be reverted in the future if future evolutions require more flexibility.