EVM Kernel: Replace the V45 migration to match Farfadet r1
* What

The V45 migration logic is updated to accurately reflect the intended
state of the Farfadet r1 deployment by dynamically computing the hash
of the injected FA deposit transaction rather than relying on a
hardcoded value.

* Why

The commit that was merged on master was an outdated version of
the intended change.

See commit d748ae500d2a9d7ac381053d436e8992f8d731bd for reference.

* How

In `etherlink/kernel_latest/kernel/src/migration.rs`, the logic inside
the migration function is refactored. The `FaDeposit` struct is now
instantiated as a named variable `fa_deposit` prior to the transaction
creation. The transaction hash is no longer defined as a hardcoded byte
array; instead, it is derived programmatically by calling
`fa_deposit.hash` (using a zeroed 20-byte key). This computed `tx_hash`
is then used to construct the final `Transaction` object that is saved
to the delayed inbox.

This reflects the migration that has been merged in the `etherlink-farfadet-r1`
branch.